pyproject.toml
[tool.poetry]
name = "DocWain"
version = "1.0.0"
description = "An intelligent document processing and QA system"
authors = ["Muthu Subramanian muthu.subramanian@dhsit.co.uk"]
license = "MIT"
readme = "README.md"
packages = [{include = "docwain"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "^1.36.0"
langchain-community = "^0.2.6"
langchain-experimental = "^0.0.62"
langchain-groq = "^0.1.6"
langchain-qdrant = "^0.1.1"
flashrank = "^0.2.8"
qdrant-client = "^1.10.1"
pypdfium2 = "^4.30.0"
fastembed = "^0.3.3"
python-dotenv = "^1.0.1"
python-docx = "^0.8.11"
pandas = "^2.0.0"
openpyxl = "^3.1.2"
tqdm = "^4.65.0"
unstructured = "^0.10.0"
boto3 = "^1.34.0"
aiofiles = "^23.2.1"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"